refactor(webui): split app modules and clean artifacts
