+++ pages/dashboard/index.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Telegram Forwarder</title>
-    <link rel="stylesheet" href="./assets/style.css?v=3a701ff489" />
-    <script src="./assets/js/gsap.min.js?v=3a701ff489"></script>
+    <link rel="stylesheet" href="./assets/style.css?v=3cb1973a9f" />
+    <script src="./assets/js/gsap.min.js?v=3cb1973a9f"></script>
   </head>
   <body>
     <div class="bg-aurora" aria-hidden="true"><i></i><i></i></div>
@@ -403,6 +403,6 @@
 
     <div class="toast" id="toast"></div>
     <script src="/api/plugin/page/bridge-sdk.js"></script>
-    <script src="./assets/app.js?v=3a701ff489" type="module" defer></script>
+    <script src="./assets/app.js?v=3cb1973a9f" type="module" defer></script>
   </body>
 </html>
